requirements.txt
atomicwrites==1.4.1; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.4.0"
attrs==22.1.0; python_version >= "3.7"
certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.0; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0"
flake8==4.0.1; python_version >= "3.6"
grpcio-tools==1.48.0; python_version >= "3.6"
grpcio==1.48.0; python_version >= "3.6"
idna==3.3; python_version >= "3.7" and python_version < "4"
iniconfig==1.1.1; python_version >= "3.7"
mccabe==0.6.1; python_version >= "3.6"
packaging==21.3; python_version >= "3.7"
pluggy==1.0.0; python_version >= "3.7"
protobuf==3.20.1; python_version >= "3.7"
py==1.11.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
pycodestyle==2.8.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pyflakes==2.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pyhamcrest==2.0.3; python_version >= "3.5"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.7"
pytest==7.1.2; python_version >= "3.7"
requests==2.28.1; python_version >= "3.7" and python_version < "4"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
tomli==2.0.1; python_version >= "3.7"
urllib3==1.26.11; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"

pytest~=7.1.2
requests~=2.28.1
 sqlalchemy~=1.4.40